fix: iOS image rendering and consolidate content part dispatch to commonMain
- Configure Coil ImageLoader with authenticated Ktor HttpClient on iOS
- Extract ContentPartRenderer sub-composables and tool call dispatch into shared commonMain (SharedContentParts.kt, ToolCallParsing.kt)
- Add missing iOS handlers: AGENT_UPDATE, VIDEO_URL null fallback
- Fix error text fallback inconsistency across platforms
- Add memory cache to iOS ImageLoader
