+++ shared/src/commonMain/kotlin/com/garfiec/librechat/shared/app/LibreChatApp.kt
@@ -4,10 +4,17 @@ import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import coil3.ImageLoader
+import coil3.compose.setSingletonImageLoaderFactory
+import coil3.memory.MemoryCache
+import coil3.network.ktor3.KtorNetworkFetcherFactory
+import coil3.request.crossfade
 import com.garfiec.librechat.core.data.datastore.ThemeDataStore
 import com.garfiec.librechat.core.data.datastore.ThemeMode
 import com.garfiec.librechat.core.ui.theme.LibreChatTheme
 import com.garfiec.librechat.shared.navigation.LibreChatNavHost
+import io.ktor.client.HttpClient
 import org.koin.compose.koinInject
 
 /**
@@ -22,6 +29,24 @@ import org.koin.compose.koinInject
  */
 @Composable
 fun LibreChatApp() {
+    val httpClient = koinInject<HttpClient>()
+    val imageLoaderFactory = remember(httpClient) {
+        { context: coil3.PlatformContext ->
+            ImageLoader.Builder(context)
+                .components {
+                    add(KtorNetworkFetcherFactory(httpClient))
+                }
+                .memoryCache {
+                    MemoryCache.Builder()
+                        .maxSizePercent(context, 0.25)
+                        .build()
+                }
+                .crossfade(true)
+                .build()
+        }
+    }
+    setSingletonImageLoaderFactory(imageLoaderFactory)
+
     val themeDataStore = koinInject<ThemeDataStore>()
     val themeMode by themeDataStore.themeMode.collectAsState(initial = themeDataStore.initialThemeMode)
     val systemDark = isSystemInDarkTheme()
